refactor: rename ichiarom module to aromaticity and add Doxygen
Rename ichiarom.{c,h} -> aromaticity.{c,h} (and test_ichiarom.cpp ->
test_aromaticity.cpp), update the include guard, includes, and all three
build systems. Replace the terse per-function comments with Doxygen blocks.
No behavior change.
