+++ INCHI-1-SRC/INCHI_API/demos/inchi_main/gcc/makefile
@@ -205,7 +205,7 @@ ichimap1.o	\
 ichimap2.o	\
 ichimap4.o	\
 ichinorm.o	\
-ichiarom.o	\
+aromaticity.o	\
 ichiparm.o	\
 ichiprt1.o	\
 ichiprt2.o	\
