+++ INCHI-1-SRC/INCHI_API/demos/mol2inchi/gcc/makefile
@@ -207,7 +207,7 @@ ichimap1.o	\
 ichimap2.o	\
 ichimap4.o	\
 ichinorm.o	\
-ichiarom.o	\
+aromaticity.o	\
 ichiparm.o	\
 ichiprt1.o	\
 ichiprt2.o	\
