+++ INCHI-1-SRC/INCHI_API/demos/test_ixa/gcc/makefile
@@ -203,7 +203,7 @@ ichimap1.o	\
 ichimap2.o	\
 ichimap4.o	\
 ichinorm.o	\
-ichiarom.o	\
+aromaticity.o	\
 ichiparm.o	\
 ichiprt1.o	\
 ichiprt2.o	\
