+++ INCHI-1-SRC/INCHI_API/libinchi/gcc/makefile
@@ -204,7 +204,7 @@ ichimap1.o	\
 ichimap2.o	\
 ichimap4.o	\
 ichinorm.o	\
-ichiarom.o	\
+aromaticity.o	\
 ichiparm.o	\
 ichiprt1.o	\
 ichiprt2.o	\
