+++ INCHI-1-SRC/INCHI_API/libinchi/src/CMakeLists.txt
@@ -82,8 +82,8 @@ target_sources(libinchi PRIVATE
 	${P_BASE}/ichimap4.c
 	${P_BASE}/ichinorm.c
 	${P_BASE}/ichinorm.h
-	${P_BASE}/ichiarom.c
-	${P_BASE}/ichiarom.h
+	${P_BASE}/aromaticity.c
+	${P_BASE}/aromaticity.h
 	${P_BASE}/ichiparm.c
 	${P_BASE}/ichiprt1.c
 	${P_BASE}/ichiprt2.c
